+++ lib/twilio-ruby/rest/media/v1/player_streamer.rb
@@ -36,7 +36,8 @@ module Twilio
           #   PlayerStreamer can run before automatically ends. The default value is 300
           #   seconds, and the maximum value is 90000 seconds. Once this maximum duration is
           #   reached, Twilio will end the PlayerStreamer, regardless of whether media is
-          #   still streaming.
+          #   still streaming. **Note: this feature has not yet been enabled but customers are
+          #   advised to explicitly set it on all their new PlayerStreamer resources.**
           # @return [PlayerStreamerInstance] Created PlayerStreamerInstance
           def create(video: :unset, status_callback: :unset, status_callback_method: :unset, max_duration: :unset)
             data = Twilio::Values.of({