+++ spec/integration/verify/v2/service/verification_check_spec.rb
@@ -38,6 +38,7 @@ describe 'VerificationCheck' do
           "valid": true,
           "amount": null,
           "payee": null,
+          "sna_attempts_error_codes": [],
           "date_created": "2015-07-30T20:00:00Z",
           "date_updated": "2015-07-30T20:00:00Z"
       }
@@ -64,6 +65,7 @@ describe 'VerificationCheck' do
           "valid": true,
           "amount": null,
           "payee": null,
+          "sna_attempts_error_codes": [],
           "date_created": "2020-01-30T20:00:00Z",
           "date_updated": "2020-01-30T20:00:00Z"
       }
@@ -90,6 +92,12 @@ describe 'VerificationCheck' do
           "valid": true,
           "amount": null,
           "payee": null,
+          "sna_attempts_error_codes": [
+              {
+                  "attempt_sid": "VLaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                  "code": 60001
+              }
+          ],
           "date_created": "2015-07-30T20:00:00Z",
           "date_updated": "2015-07-30T20:00:00Z"
       }